create policies

See merge request lbaw/lbaw2122/lbaw2181!16

+++ app/Http/Controllers/TasksController.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Tasks;
+use App\Models\Task;
 use App\Models\Project;
-use App\Models\UserAssigns;
+use App\Models\UserAssign;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class TasksController extends Controller
 {
@@ -24,6 +25,7 @@ class TasksController extends Controller
 
     public function showTaskForm(int $id)
     {
+        Gate::authorize('manager',Project::find($id));
         $notifications = NotificationsController::getNotifications(Auth::id());
 
         $users = ProjectUsersController::getProjectUsers($id);
@@ -41,6 +43,7 @@ class TasksController extends Controller
 
     public function showTask(int $project_id, int $id)
     {
+        Gate::authorize('show',Task::find($id));
         $notifications = NotificationsController::getNotifications(Auth::id());
 
         $users = ProjectUsersController::getProjectUsers($project_id);
@@ -49,54 +52,58 @@ class TasksController extends Controller
                             ->leftjoin('users', 'users.id', '=', 'userassigns.user_id')
                             ->where('task_id', '=', $id)
                             ->get(['userassigns.user_id','users.username']);
-        
+
         $user_assigned = json_decode($user_assigned, true);
 
         return view('pages.task',['project' => Project::find($project_id),
-                                    'task' => Tasks::find($id), 
-                                    'notifications' => $notifications, 
-                                    'users' => $users, 
+                                    'task' => Task::find($id),
+                                    'notifications' => $notifications,
+                                    'users' => $users,
                                     'user_assigned' => $user_assigned]);
     }
 
 
     public function updateTask(int $project_id, int $id, Request $request) {
+
+
         $notifications = NotificationsController::getNotifications(Auth::id());
 
         switch ($request->input('action'))
         {
             case 'update':
+                Gate::authorize('inProject',Project::find($project_id));
                 $validator = $request->validate($this->validator());
 
-                $task = Tasks::find($id);
+                $task = Task::find($id);
                 $task->name = $request->name;
                 $task->description = $request->description;
                 $task->due_date = $request->due_date;
                 $task->tag = $request->tag;
                 $task->save();
-                
+
                 if ($request->user_id == -1){
-                    $user_assigned = UserAssigns::where('task_id', '=', $id)
+                    $user_assigned = UserAssign::where('task_id', '=', $id)
                                             ->delete(['user_id' =>$request->user_id]);
                     break;
                 }
 
-                $user_assigned = UserAssigns::where('task_id', '=', $id)
+                $user_assigned = UserAssign::where('task_id', '=', $id)
                                             ->update(['user_id' =>$request->user_id]);
 
                 $user_assigned = json_decode($user_assigned, true);
-                
+
                 if (!$user_assigned){
-                    $user_assign = new UserAssigns;
+                    $user_assign = new UserAssign;
                     $user_assign->task_id = $task->id;
                     $user_assign->user_id = $request->user_id;
                     $user_assign->save();
                 }
-                
+
                 break;
 
             case 'delete':
-                $task=Tasks::find($id);
+                Gate::authorize('manager',Project::find($project_id));
+                $task=Task::find($id);
                 $task->delete(); //returns true/false
                 break;
         }
@@ -106,6 +113,7 @@ class TasksController extends Controller
 
     public function showTasks($project_id)
     {
+        Gate::authorize('inProject',Project::find($project_id));
         $notifications = NotificationsController::getNotifications(Auth::id());
 
         $my_TASKS = DB::table('tasks')
@@ -166,10 +174,10 @@ class TasksController extends Controller
      */
     protected function create(Request $request)
     {
+        Gate::authorize('manager',Project::find($request->project_id));
         $notifications = NotificationsController::getNotifications(Auth::id());
         $validator = $request->validate($this->validator());
-
-        $task = new Tasks;
+        $task = new Task;
         $task->name = $request->name;
         $task->description = $request->description;
         $task->due_date = $request->due_date;
@@ -180,11 +188,16 @@ class TasksController extends Controller
         $task->created_at = Carbon::now();
         $task->save();
 
+        $user_assign = new UserAssign;
+        $user_assign->task_id = $task->id;
+        $user_assign->user_id = $request->user_id;
+        $user_assign->save();
+
         return redirect()->action([TasksController::class,'showTasks'], ['id'=> $task->project_id]);
     }
 /*
     static function getProjectTasks($project_id) {
-        return (new Tasks())->where('project_id','=',$project_id)->orderBy('due_date');
+        return (new Task())->where('project_id','=',$project_id)->orderBy('due_date');
     }
 */
     public function searchProjectTasks(int $project_id, Request $request)
